pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "baymax-ai"
version = "0.1.0"
description = "Multi-agent AI career platform for Pakistani CS students"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Baymax AI Team", email = "dev@baymaxai.pk"},
]
keywords = [
    "ai",
    "career",
    "resume",
    "interview",
    "job-search",
    "groq",
    "langchain",
    "crewai",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Education",
    "Topic :: Office/Business",
]

dependencies = [
    "groq>=0.9.0",
    "langchain>=1.0.0",
    "langchain-groq>=1.0.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "langgraph>=0.2.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "pypdf2>=3.0.0",
    "pypdf>=4.0.0",
    "requests>=2.32.0",
    "httpx>=0.27.0",
    "streamlit>=1.36.0",
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.0",
    "pydantic>=2.7.0",
    "tiktoken>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pylint>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/baymax-ai"
Documentation = "https://github.com/yourusername/baymax-ai/tree/main/docs"
Repository = "https://github.com/yourusername/baymax-ai.git"
"Bug Tracker" = "https://github.com/yourusername/baymax-ai/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["agents", "tools", "rag"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_mode = 3
include_trailing_comma = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --strict-markers --tb=short"
